# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-Library System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: campus
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: course
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
        - name: section
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullname, email, phone, address, campus, role, createdAt]
      properties:
        id:
          type: integer
        fullname:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        campus:
          type: string
        role:
          type: string
        studentNumber:
          type: ["string", "null"]
        course:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        isApproved:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        fullname:
          type: string
        phone:
          type: string
        address:
          type: string
        campus:
          type: string
        isApproved:
          type: boolean
        course:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
...