# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Library System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  fullname?: string;
  phone?: string;
  address?: string;
  campus?: string;
  isApproved?: boolean;
  /** @nullable */
  course?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  section?: string | null;
}